code/week-3-backend/requirements.txt
flask
flask-cors
boto3
pandas
python-dotenv
